requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Werkzeug==2.3.7
pandas==2.1.1
openpyxl==3.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
